Corrigindo Flooding

+++ src/routing/Flooding.cc
@@ -98,7 +98,7 @@ void Flooding::finish() {
         std::ofstream log;
         std::ostringstream o;
 
-        o << "./results/" << par("log_traffic") << "-" << par("log_replication") << "-receiver-" << myId;
+        o << "./results/" << par("log_traffic").longValue() << "-" << par("log_replication").longValue() << "-receiver-" << myId;
         log.open(o.str().c_str());
 
         for (std::map<int, MessageInfoEntry*>::iterator i = messagesRcvd.begin(); i != messagesRcvd.end(); i++) {
